refactor(website): split projects-db into focused modules (S1 fix)
Split projects-db.ts (803 lines) into 4 focused modules to comply with S1 quality gate (≤600 lines per .ts file):
- projects-db.ts: 433 lines (core Projects/SubProject/ProjectTask CRUD)
- project-attachments-db.ts: 67 lines (ProjectAttachment operations)
- project-portal-db.ts: 161 lines (Portal access, customer functions)
- project-export-db.ts: 155 lines (CSV export, meeting relations)

Also fixed S2 cycle (website-db.ts → projects-db.ts) by removing re-export.
All Projects imports now point directly to the modules they need.

[T001359]
